% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{flow_centrality}
\alias{flow_centrality}
\title{Determine dendrite and axon by calculating flow centrality}
\usage{
flow_centrality(
  x,
  mode = c("sum", "centrifugal", "centripetal"),
  polypre = TRUE,
  primary.dendrite = 0.9,
  bending.flow = FALSE,
  split = c("postsynapses", "presynapses", "distance")
)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object}

\item{mode}{type of flow centrality to calculate. There are three flavors:
(1) centrifugal, which counts paths from proximal inputs to distal outputs;
(2) centripetal, which counts paths from distal inputs to proximal outputs;
and (3) the sum of both.}

\item{polypre}{whether to consider the number of presynapses as a multiple of
the numbers of connections each makes}

\item{primary.dendrite}{whether to try to assign nodes to a 'primary
dendrite'. Defaults to considering nodes of 0.9*maximal flow centrality.
Assigning to NULL will prevent generating this compartment.}

\item{bending.flow}{we may need to add the 'bending flow' to all the branchpoints if looking at centripetal flow centrality}

\item{split}{the algorithm will assign two main neurite compartments, which as per SWC format will be indicates as either axon (Label =2)
or dendrite (Label = 3) in the returned objects, at neuron$d$Label.
This assignment can be based which compartment contains the most postsynapses ("postsynapses") or presynapses ("presynapses"),
or the Euclidean distance of its first branch point from the primary branch point (i.e. the first branch point from the soma) ("distance").}
}
\value{
the neuron or neuron list object inputted, with centripetal flow
  centrality information added to neuron$d and a segregation index score.
}
\description{
implementation of the algorithm for calculating flow
  centralities from Schneider-Mizell et al. (2016). Can be used on \code{nat::neuronlist}
  objects read using \code{\link{neuprint_read_neurons}}.
}
\details{
From Schneider-Mizell et al. (2016): "We use flow centrality for
  four purposes. First, to split an arbor into axon and dendrite at the
  maximum centrifugal SFC, which is a preliminary step for computing the
  segregation index, for expressing all kinds of connectivity edges (e.g.
  axo-axonic, dendro-dendritic) in the wiring diagram, or for rendering the
  arbor in 3d with differently colored regions. Second, to quantitatively
  estimate the cable distance between the axon terminals and dendritic arbor
  by measuring the amount of cable with the maximum centrifugal SFC value.
  Third, to measure the cable length of the main dendritic shafts using
  centripetal SFC, which applies only to insect neurons with at least one
  output syn- apse in their dendritic arbor. And fourth, to weigh the color
  of each skeleton node in a 3d view, providing a characteristic signature of
  the arbor that enables subjective evaluation of its identity."
}
\examples{
\donttest{
\dontrun{
# Get neuron
neuron = neuprint_read_neuron("818983130")

# Split neuron
neuron.flow = flow_centrality(neuron)

# Plot split
dendrites.v = subset(rownames(neuron.flow$d), neuron.flow$d$Label == 3)
axon.v = subset(rownames(neuron.flow$d), neuron.flow$d$Label == 2)
dendrites = nat::prune_vertices(neuron.flow, verticestoprune = dendrites.v, invert = TRUE)
axon = nat::prune_vertices(neuron.flow, verticestoprune = axon.v, invert = TRUE)
rgl::plot3d(dendrites, col = "cyan", lwd = 2)
rgl::plot3d(axon, col = "firebrick", lwd = 2)
rgl::plot3d(neuron, col = "grey30",lwd=2)
}}
}
\references{
Schneider-Mizell, C. M., Gerhard, S., Longair, M., Kazimiers, T.,
  Li, F., Zwart, M. F., … Cardona, A. (2015). Quantitative neuroanatomy for
  connectomics in Drosophila. bioRxiv, 026617. http://doi.org/10.1101/026617
}
\seealso{
\code{\link{neuprint_read_neurons}} \code{\link{neuprint_get_synapses}} \code{\link{neuprint_locate_soma}}
}
