% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{heal_skeleton}
\alias{heal_skeleton}
\title{Heal a fragmented skeleton for a neuron}
\usage{
heal_skeleton(x, ...)
}
\arguments{
\item{x}{either an object of class neuron, or neuronlist}

\item{...}{methods passed to \code{neuprint_login}}
}
\value{
a cohesive SWC like data frame, or a cohesive neuron/neuronlist object as dictated used by the \code{nat} and \code{rcatmaid} packages
}
\description{
Mend breaks in a skeleton for a neuron, predicting merge sites using the minimum spanning tree method, utilising \code{igraph::mst}
}
\seealso{
\code{\link{neuprint_get_synapses}}, \code{\link{neuprint_read_neurons}}
}
