% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{neuprint_connection_table}
\alias{neuprint_connection_table}
\title{Get the upstream and downstream connectivity of a neuron}
\usage{
neuprint_connection_table(
  bodyids,
  partners = c("inputs", "outputs"),
  prepost = c("PRE", "POST"),
  roi = NULL,
  by.roi = FALSE,
  threshold = 1L,
  summary = FALSE,
  details = FALSE,
  superLevel = FALSE,
  progress = FALSE,
  dataset = NULL,
  chunk = TRUE,
  all_segments = FALSE,
  conn = NULL,
  ...
)
}
\arguments{
\item{bodyids}{the body IDs for neurons/segments (bodies) you wish to query.
This can be in any form understood by \code{\link{neuprint_ids}}.}

\item{partners}{\code{inputs} looks for upstream inputs (presynaptic) whereas
\code{outputs} looks for downstream outputs (postsynaptic) to the given
\code{bodyids}.}

\item{prepost}{\code{PRE}: look for partners presynaptic (i.e upstream
inputs) or \code{POST}: postsynaptic (downstream outputs) to the given
\code{bodyids}. NB this is redundant to the \code{partners} argument and
you should only use one.}

\item{roi}{a single ROI. Use \code{neuprint_ROIs} to see what is available.}

\item{by.roi}{logical, whether or not to break neurons' connectivity down by
region of interest (ROI)}

\item{threshold}{Only return partners >= to an integer value. Default of 1
returns all partners. This threshold will be applied to the ROI weight when
the \code{roi} argument is specified, otherwise to the whole neuron.}

\item{summary}{When \code{TRUE} and more than one query neuron is given,
summarises connectivity grouped by partner.}

\item{details}{When \code{TRUE} returns adds a name and type column for
partners. Alternatively a character vector of column names can be provided
(but see \emph{details} for details about \code{details}). Default
\code{details=FALSE}.}

\item{superLevel}{When \code{by.roi=TRUE}, should we look at low-level ROIs
(\code{superLevel=FALSE}) or only super-level ROIs
(\code{superLevel=TRUE}). A super-level ROIs can contain multiple
lower-level ROIs. If set to `NULL`, both are returned.}

\item{progress}{default FALSE. If TRUE, the API is called separately for each
neuron and you can assess its progress, if an error is thrown by any one
\code{bodyid}, that \code{bodyid} is ignored}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link{neuprint_login}} for details.}

\item{chunk}{A logical specifying whether to split the query into multiple
chunks or an integer specifying the size of those chunks (which defaults to
20 when \code{chunk=TRUE}).}

\item{all_segments}{if TRUE, all bodies are considered, if FALSE, only
'Neurons', i.e. bodies with a status roughly traced status.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}

\item{...}{arguments passed to \code{nat::nlapply} and downstream helper
functions such as \code{neuprint_read_neuron}.}
}
\value{
a data frame with one row for each unique combination of query
  neuron, partner neuron and (where specified) ROI. \itemize{

  \item bodyid query neuron identifier

  \item partner neuron identifier

  \item prepost 0 for upstream/input partners; 1 for downstream/output
  partners.

  \item weight total number of connections between the query and partner
  neuron.

  }. When \code{by.roi=TRUE}

  \itemize{

  \item roi the name of the ROI

  \item ROIweight the number of connections within that ROI

  }

  Note that when \code{by.roi=TRUE} there may be multiple rows for each
  unique pair of neurons \bold{and} connections may appear in more than one
  row when some ROIs are contained within larger ROIs. For example the
  \code{CA(R)} is contained within the \code{MB(R)} ROI. Therefore you cannot
  sum the \code{ROIweight} to match the total connection strength (which is
  available as the \code{weight} column).
}
\description{
Get the upstream and downstream connectivity of a body,
  restricted to within an ROI if specified
}
\details{
Note that by default neuprint_connection_table only returns upstream
  or downstream connections with objects that were large enough to be
  designated as "Neurons". Smaller objects (Segments) can be included when
  \code{all_segments=TRUE}. When this is done then the total counts will
  match what is reported by \code{\link{neuprint_get_meta}}.

  The user must ensure that any columns named in the \code{details} argument
  are correct (non-existent columns will be filled with \code{NA}). For
  historical reasons the column displayed by neuprintr as \code{name} must be
  referred to as \code{instance} in this argument (but will be renamed to
  \code{name}). \href{https://github.com/natverse/neuprintr/issues}{File an
  issue} if this bothers you and I will try to do something smarter.
}
\examples{
\donttest{
## Unitary connection strengths between two chosen neurons,
## and their downstream targets
c1 = neuprint_connection_table(c(818983130, 1796818119), prepost = "POST")
head(c1)

# query of regex against cell type
# summarised per partner with additional details column
c1s = neuprint_connection_table("/DA2.*lPN", partners='out', summary=TRUE, details=TRUE)
head(c1s)
# Kenyon cells typically receive fewer multiple inputs than other partners
table(n=c1s$n, KC=grepl("^KC", c1s$type))

## The same connections broken down by ROI
c2 = neuprint_connection_table(c(818983130, 1796818119), prepost = "POST",
                               by.roi = TRUE)
head(c2)

## The same connections broken down by super-level ROIs only
c3 = neuprint_connection_table(c(818983130, 1796818119), prepost = "POST",
                               by.roi = TRUE, superLevel = TRUE)
nrow(c3)
nrow(c2)

## Find only the connections within a specific ROI
c4 = neuprint_connection_table(c(818983130, 1796818119), prepost = "POST",
                               by.roi = TRUE, roi = "LH(R)")

}

\donttest{
# compare pre/post/upstream/downstream when all_segments=TRUE or FALSE
neuprint_get_meta('/DP1m.*vPN')
sum(neuprint_connection_table('/DP1m.*vPN', partners = 'in')$weight)
sum(neuprint_connection_table('/DP1m.*vPN', partners = 'in', all_segments = TRUE)$weight)
sum(neuprint_connection_table('/DP1m.*vPN', partners = 'out')$weight)
sum(neuprint_connection_table('/DP1m.*vPN', partners = 'out', all_segments = TRUE)$weight)
}
}
\seealso{
\code{\link{neuprint_fetch_custom}},
  \code{\link{neuprint_simple_connectivity}},
  \code{\link{neuprint_common_connectivity}}, \code{\link{neuprint_ROIs}}
}
