% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.R
\name{neuprint_cosine_matrix}
\alias{neuprint_cosine_matrix}
\title{Calculate a cosine similarity matrix for neuprint neurons}
\usage{
neuprint_cosine_matrix(
  ids,
  ...,
  threshold = 5,
  partners = c("outputs", "inputs"),
  group = FALSE,
  details = NULL,
  conn = NULL
)
}
\arguments{
\item{ids}{Passed to \code{\link{neuprint_ids}}}

\item{...}{Optional filter expression defining which partners to include}

\item{threshold}{An integer threshold (connections >= this will be returned)}

\item{partners}{Whether to cluster based on connections to input or output
partner neurons (default both).}

\item{group}{Whether to group by cell \code{type} or another named column.}

\item{details}{Optional character vector naming metadata columns to fetch for
partner neurons.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}
}
\value{
matrix or list of two matrices (input and output)
}
\description{
Calculate a cosine similarity matrix for neuprint neurons
}
\examples{
\dontrun{
neuprint_cosine_matrix("/lLN2.+", grepl("PN",type), partners='in')
}
}
\seealso{
\code{\link{neuprint_cosine_plot}}
}
