% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.R
\name{neuprint_cosine_plot}
\alias{neuprint_cosine_plot}
\title{Plot cosine clustering of neuprint/hemibrain neurons}
\usage{
neuprint_cosine_plot(
  x,
  partners = c("inputs", "outputs"),
  threshold = 5,
  method = c("ward.D", "single", "complete", "average", "mcquitty", "median", "centroid",
    "ward.D2"),
  labRow = "{type}",
  nas = c("zero", "drop"),
  conn = NULL,
  ...
)
}
\arguments{
\item{x}{Query ids. May including searches in the style of
\code{\link{neuprint_ids}}}

\item{partners}{Whether to cluster based on connections to input or output
partner neurons (default both).}

\item{threshold}{An integer threshold (connections >= this will be returned)}

\item{method}{The cluster method to use (see \code{\link{hclust}})}

\item{labRow}{A vector defining the row labels. If it is a character vector
length 1 containing any curly braces it will be interpreted as a
\code{\link{glue}} string for interpolation, using the data.frame returned
by \code{\link{neuprint_get_meta}} as a source of information.}

\item{nas}{What to do with entries that have NAs. Default is to set them to 0
similarity.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}

\item{...}{Additional arguments passed to \code{\link{heatmap}}}
}
\value{
The result of \code{\link{heatmap}} invisibly including the row and
  column dendrograms.
}
\description{
Plot cosine clustering of neuprint/hemibrain neurons
}
\examples{
\donttest{
# Considering all partners
neuprint_cosine_matrix("/lLN2.+", partners='in') \%>\% neuprint_cosine_plot()
# excluding the PN partners from the cosine distance score
neuprint_cosine_matrix("/lLN2.+", !grepl("PN",type), partners='in') \%>\% neuprint_cosine_plot()
# just use PN partners for cosine distance score
neuprint_cosine_matrix("/lLN2.+", grepl("PN",type), partners='in') \%>\% neuprint_cosine_plot()
}
}
