% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{neuprint_dump}
\alias{neuprint_dump}
\title{Download data from neuprint for specified bodyids / a specified ROI}
\usage{
neuprint_dump(
  dir,
  bodyids = NULL,
  roi = NULL,
  preprocess = NULL,
  connectivity = TRUE,
  volumes = TRUE,
  meta = TRUE,
  nat = TRUE,
  drvid = FALSE,
  soma = TRUE,
  heal = TRUE,
  connectors = TRUE,
  all_segments = FALSE,
  resample = FALSE,
  scale = 4,
  voxel.thresh = 1e+07,
  dataset = NULL,
  conn = NULL,
  OmitFailures = TRUE,
  ...
)
}
\arguments{
\item{dir}{the directory to which to save the dump}

\item{bodyids}{the body IDs for neurons/segments (bodies) you wish to query.
This can be in any form understood by \code{\link{neuprint_ids}}.}

\item{roi}{a single ROI. Use \code{neuprint_ROIs} to see what is available.}

\item{preprocess}{a function that can be applied to a nat::neuronlist object, to be called on neurons once they are read from
the neuprint server using \code{neuprint_read_neurons}}

\item{connectivity}{whether or not to include connectivity information in the dump, i.e. an adjacency matrix between bodyids}

\item{volumes}{whether or not to include neuron volume information in the dump, i.e. voxels. Currently only  works by talking to a DVID server using
the package drvid}

\item{meta}{whether or not to fetch a meta data for the given bodyids, using
\code{neuprint_get_meta}}

\item{nat}{whether or not to read neurons as \code{nat::\link[nat]{neuron}}
objects (TRUE) or get SWC data frame (FALSE)}

\item{drvid}{whether or not to use \code{drvid::read.neuron.dvid} rather than
a cypher post request to \code{neuprint_fetch_custom}, in order to read a
neuron. This might be faster, and this might also enable access to
skeletons on an underlying DVID database that have not yet been ported to
neuprint.}

\item{soma}{whether or not to fetch a possible soma location for the given
bodyids, using \code{neuprint_locate_soma}}

\item{heal}{whether or not to heal a fragmented skeleton using a minimum
spanning tree, via \code{nat::stitch_neurons_mst}}

\item{connectors}{whether or not to add synapse data to the retrieved
skeletons in the format used by the \code{rcatmaid} package, for easy use
with \code{rcatmaid} or \code{catnat} functions. This can be done for
synapse-less skeletons using \code{neuprint_assign_connectors}}

\item{all_segments}{if TRUE, all bodies are considered, if FALSE, only
'Neurons', i.e. bodies with a status roughly traced status.}

\item{resample}{if a number, the neuron is resampled using
\code{nat::resample}, stepsize = resample. If 0 or FALSE (default), no
resampling occurs.}

\item{scale}{Resolution of sparse volume starting with 0 where each level
beyond 0 has 1/2 resolution of previous level. "coarse" will return the
volume in block coordinates.}

\item{voxel.thresh}{the size threshold, in number of voxels, a neuron/segment must exceed, to be included in the dump, if read from an ROI}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link{neuprint_login}} for details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}

\item{OmitFailures}{Whether to omit neurons for which \code{FUN} gives an
error. The default value (\code{NA}) will result in nlapply stopping with
an error message the moment there is an error. For other values, see
details.}

\item{...}{additional arguments passed to helper functions including
\code{neuprint_bodies_in_ROI}, \code{neuprint_read_neurons},
\code{neuprint_get_adjacency_matrix}, \code{neuprint_connection_table},
and \code{drvid::dv_get_voxels}.}
}
\description{
Download neuron morphology and connectivity data to a specified directory as .csv and .rda files
}
\seealso{
\code{\link{neuprint_get_synapses}}, \code{\link{neuprint_read_neurons}}
}
