% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R, R/roi.R
\name{neuprint_connection_table}
\alias{neuprint_connection_table}
\alias{neuprint_find_neurons}
\alias{neuprint_bodies_in_ROI}
\title{Get the upstream and downstream connectivity of a neuron}
\usage{
neuprint_connection_table(bodyids, prepost = c("PRE", "POST"),
  roi = NULL, progress = FALSE, dataset = NULL,
  all_segments = TRUE, conn = NULL, ...)

neuprint_find_neurons(input_ROIs, output_ROIs = NULL, statuses = NULL,
  dataset = NULL, conn = NULL, all_segments = TRUE, ...)

neuprint_bodies_in_ROI(roi, dataset = NULL, all_segments = TRUE,
  conn = NULL, ...)
}
\arguments{
\item{bodyids}{the body IDs for neurons/segments (bodies) you wish to query}

\item{prepost}{whether to look for partners presynaptic to postsynaptic to the given bodyids}

\item{roi}{a single ROI. Use \code{neuprint_ROIs} to see what is available.}

\item{progress}{default FALSE. If TRUE, the API is called separately for each neuron and yuo can asses its progress, if an error is thrown by any one bodyid, that bodyid is ignored}

\item{dataset}{optional, a dataset you want to query. If NULL, the default specified by your R environ file is used. See \code{neuprint_login} for details.}

\item{all_segments}{if TRUE, all bodies are considered, if FALSE, only 'Neurons', i.e. bodies with a status roughly traced status.}

\item{conn}{optional, a neuprintr connection object, which also specifies the neuPrint server see \code{?neuprint_login}.
If NULL, your defaults set in your R.profile or R.environ are used.}

\item{...}{methods passed to \code{neuprint_login}}

\item{input_ROIs}{a vector of input ROIs. Use \code{neuprint_ROIs} to see what is available.}

\item{output_ROIs}{a vector of input ROIs. Use \code{neuprint_ROIs} to see what is available.}

\item{statuses}{if not NULL, only bodies with the given status are considered. Statuses include:
Unimportant,0.5assign,Leaves,Prelim Roughly Traced, Anchor, Orphan.}
}
\value{
a data frame giving partners within an ROI, the connection strength for weights to or from that partner, and the direction, for the given bodyid

a n x n matrix, where the rows are input neurons and the columns are their targets
}
\description{
Get the upstream and downstream connectivity of a body, restricted to within an ROI if specified

Find bodies that input/output within certain ROIs (i.e. they must have synapses in these areas)
}
\seealso{
\code{\link{neuprint_fetch_custom}}, \code{\link{neuprint_simple_connectivity}}, \code{\link{neuprint_common_connectivity}}

\code{\link{neuprint_fetch_custom}}, \code{\link{neuprint_simple_connectivity}}, \code{\link{neuprint_common_connectivity}}
}
