% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{neuprint_get_fields}
\alias{neuprint_get_fields}
\title{Get available metadata fields for Neuron nodes}
\usage{
neuprint_get_fields(
  possibleFields = c("bodyId", "pre", "post", "upstream", "downstream", "status",
    "statusLabel", "cropped", "instance", "name", "size", "type", "cellBodyFiber",
    "somaLocation", "somaRadius"),
  limit = 20,
  negateFields = FALSE,
  dataset = NULL,
  conn = NULL,
  ...
)
}
\arguments{
\item{possibleFields}{: field names to choose from}

\item{limit}{: Max length of the server's response (used to speed up the query/adapt to different scenarios)}

\item{negateFields}{: Whether to include (FALSE, the default) or exclude \code{possibleFields}}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link{neuprint_login}} for details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}

\item{...}{methods passed to \code{neuprint_fetch_custom}}
}
\value{
a vector of available fields
}
\description{
Get available metadata fields for Neuron nodes
}
\examples{
\donttest{
neuprint_get_fields()
}
}
