% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{neuprint_get_meta}
\alias{neuprint_get_meta}
\title{Get key metadata for body (including name, type, status, size)}
\usage{
neuprint_get_meta(
  bodyids,
  dataset = NULL,
  all_segments = TRUE,
  conn = NULL,
  cache = FALSE,
  chunk = TRUE,
  progress = FALSE,
  possibleFields = NULL,
  ...
)
}
\arguments{
\item{bodyids}{the body IDs for neurons/segments (bodies) you wish to query.
This can be in any form understood by \code{\link{neuprint_ids}}.}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link{neuprint_login}} for details.}

\item{all_segments}{if TRUE, all bodies are considered, if FALSE, only
'Neurons', i.e. bodies with a status roughly traced status.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}

\item{cache}{the query to neuPrint server, so that it does not need to be
repeated. Of course you can save the results, but this may be helpful e.g.
inside a wrapper function that post-processes the results like
\code{hemibrainr::grouped_adjacency_matrix}.}

\item{chunk}{A logical specifying whether to split the query into multiple
chunks or an integer specifying the size of those chunks (which defaults to
30,000 when \code{chunk=TRUE}).}

\item{progress}{default FALSE. If TRUE, the API is called separately for each
neuron and you can assess its progress, if an error is thrown by any one
\code{bodyid}, that \code{bodyid} is ignored.}

\item{possibleFields}{passed to \code{\link{neuprint_get_fields}} when not
\code{NULL}, otherwise a default set are used.}

\item{...}{arguments passed to \code{nat::nlapply} and downstream helper
functions such as \code{neuprint_read_neuron}.}
}
\value{
a \code{data.frame} containing the neuron's \itemize{

  \item name

  \item type Cell type of the neuron

  \item status (Traced etc)

  \item statusLabel similar to \code{status} but often a bit more specific

  \item size size in voxels

  \item pre number of presynapses

  \item post number of postsynapses

  \item cropped whether the neuron is cropped by the hemibrain volume

  \item soma whether the neuron has a soma in the hemibrain volume

  \item cellBodyFiber names the tract connecting the soma to rest of neuron

  }
}
\description{
Get key metadata for body (including name, type, status, size)
}
\details{
Sometimes a \code{cellBodyFiber} will be present even when the
  \code{soma} is not, so this may be a good test of if the neuron is present
  in the volume. The \code{cellBodyFiber} should be matched to (hemi)lineage
  information .
}
\examples{
\donttest{
da2s=neuprint_search(".*DA2.*")
neuprint_get_meta(da2s$bodyid)
# or simpler
neuprint_get_meta('DA2')
}
\dontrun{
neuprint_get_meta('cropped:false')
}
}
