% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{neuprint_get_neuron_names}
\alias{neuprint_get_neuron_names}
\title{Get the name of a neuron}
\usage{
neuprint_get_neuron_names(bodyids, dataset = NULL, all_segments = TRUE,
  conn = NULL, ...)
}
\arguments{
\item{bodyids}{the body IDs for neurons/segments (bodies) you wish to query}

\item{dataset}{optional, a dataset you want to query. If NULL, the default specified by your R environ file is used. See \code{neuprint_login} for details.}

\item{all_segments}{if TRUE, all bodies are considered, if FALSE, only 'Neurons', i.e. bodies with a status roughly traced status.}

\item{conn}{optional, a neuprintr connection object, which also specifies the neuPrint server see \code{?neuprint_login}.
If NULL, your defaults set in your R.profile or R.environ are used.}

\item{...}{methods passed to \code{neuprint_login}}
}
\value{
a dataframe givng a bodyid name correspondence
}
\description{
If a bodyID has a name associated with it, fetch that name, otherwise, return NA
}
