% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{neuprint_get_paths}
\alias{neuprint_get_paths}
\title{Get a list of paths of length n between 2 neurons}
\usage{
neuprint_get_paths(
  body_pre,
  body_post,
  n,
  weightT = 5,
  roi = NULL,
  by.roi = FALSE,
  exclude.loops = TRUE,
  chunk = TRUE,
  progress = FALSE,
  dataset = NULL,
  conn = NULL,
  all_segments = FALSE,
  ...
)
}
\arguments{
\item{body_pre}{the bodyid of the neuron at the start of the path}

\item{body_post}{the bodyid of the neuron at the end of the path}

\item{n}{the length of the path. If \code{n} is a vector, paths of length
n[1] to n[2] are considered}

\item{weightT}{weight threshold}

\item{roi}{Limit the search to connections happening within a certain ROI or
set of ROIs (NULL by default)}

\item{by.roi}{Return the results by ROI. Default to FALSE}

\item{exclude.loops}{Wether or not to exclude loops
(paths containing the same node several times). Defaults to TRUE}

\item{chunk}{A logical specifying whether to split the query into multiple
chunks or an integer specifiying the size of those chunks (which defaults
to 5 when \code{chunk=TRUE}).}

\item{progress}{if TRUE, a progress bar will be shown. This may slow the data
fetching process for smaller queries. The default of
\code{progress=NULL} will only show a progress bar if the query will be
split into multiple chunks based on the \code{chunk} argument.}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link{neuprint_login}} for details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}

\item{all_segments}{if TRUE, all bodies are considered, if FALSE, only
'Neurons', i.e. bodies with a status roughly traced status.}

\item{...}{methods passed to \code{neuprint_login}}
}
\description{
Get all of the paths in the database that connect the query
  neurons with at least \code{weightT} synapses at each step
}
\examples{
\donttest{
neuprint_get_paths(c(1128092885,481121605),5813041365, n=c(1,2), weightT=20)

neuprint_get_paths(c(1128092885,481121605),5813041365, n=c(1,2),
  weightT=20,by.roi=TRUE)

neuprint_get_paths(c(1128092885,481121605),5813041365,
  n=c(1,2), weightT=20,roi=c("FB","LAL(-GA)(R)"))
}
}
\seealso{
\code{\link{neuprint_get_shortest_paths}},
  \code{\link{neuprint_common_connectivity}},
  \code{\link{neuprint_get_adjacency_matrix}}
}
