% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{neuprint_get_shortest_paths}
\alias{neuprint_get_shortest_paths}
\title{Get a list of the shortest paths between two neurons}
\usage{
neuprint_get_shortest_paths(
  body_pre,
  body_post,
  weightT = 5,
  roi = NULL,
  by.roi = FALSE,
  dataset = NULL,
  conn = NULL,
  all_segments = FALSE,
  ...
)
}
\arguments{
\item{body_pre}{the bodyid of the neuron at the start of the path}

\item{body_post}{the bodyid of the neuron at the end of the path}

\item{weightT}{weight threshold}

\item{roi}{Limit the search to connections happening within a certain ROI or
set of ROIs (NULL by default)}

\item{by.roi}{Return the results by ROI. Default to FALSE}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link{neuprint_login}} for details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}

\item{all_segments}{if TRUE, all bodies are considered, if FALSE, only
'Neurons', i.e. bodies with a status roughly traced status.}

\item{...}{methods passed to \code{neuprint_login}}
}
\description{
Get all of the shortest paths in the database that connect the
  query neurons with at least weightT synapses at each step
}
\examples{
\donttest{
neuprint_get_shortest_paths(c(1128092885,481121605),5813041365,weightT=20)
}
}
\seealso{
\code{\link{neuprint_get_paths}},
  \code{\link{neuprint_common_connectivity}},
  \code{\link{neuprint_get_adjacency_matrix}}
}
