% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{neuprint_ROI_connectivity}
\alias{neuprint_ROI_connectivity}
\title{Get the connectivity between ROIs in a neuPrint dataset}
\usage{
neuprint_ROI_connectivity(rois, dataset = NULL, conn = NULL, ...)
}
\arguments{
\item{rois}{regions of interest for a dataset}

\item{dataset}{optional, a dataset you want to query. If NULL, the default specified by your R environ file is used. See \code{neuprint_login} for details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the neuPrint server see \code{?neuprint_login}.
If NULL, your defaults set in your R.profile or R.environ are used.}

\item{...}{methods passed to \code{neuprint_login}}
}
\description{
Get the connectivity between ROIs in a neuPrint dataset
}
\seealso{
\code{\link{neuprint_simple_connectivity}}, \code{\link{neuprint_common_connectivity}}
}
