% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{neuprint_ROIs}
\alias{neuprint_ROIs}
\title{Get a vector of all the ROIs in a specified dataset}
\usage{
neuprint_ROIs(superLevel = FALSE, dataset = NULL, conn = NULL, ...)
}
\arguments{
\item{superLevel}{whether not to show 'superlevel' ROIs - ROIs composed of other ROIs.
If set to NULL, both low-level and superlevel ROIs are returned.}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link{neuprint_login}} for details.}

\item{...}{methods passed to \code{neuprint_login}}
}
\description{
Get the regions of interest (ROIs) used in a neuPrint project
}
\seealso{
\code{\link{neuprint_login}}, \code{\link{neuprint_datasets}}
}
