% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{neuprint_simple_connectivity}
\alias{neuprint_simple_connectivity}
\title{Get a list of connected partners for a neuron}
\usage{
neuprint_simple_connectivity(bodyids, prepost = c("PRE", "POST"),
  dataset = NULL, conn = NULL, ...)
}
\arguments{
\item{bodyids}{the cypher by which to make your search}

\item{prepost}{whether to look for partners presynaptic ot postsynaptic to the given bodyids}

\item{dataset}{optional, a dataset you want to query. If NULL, the default specified by your R environ file is used. See \code{neuprint_login} for details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the neuPrint server see \code{?neuprint_login}.
If NULL, your defaults set in your R.profile or R.environ are used.}

\item{...}{methods passed to \code{neuprint_login}}
}
\value{
a n x m matrix where n correspond to the neurons that connect to m body ids
}
\description{
Get all of the neurons in the database that connect to the query neurons, either upstream or downstream of them
}
\seealso{
\code{\link{neuprint_common_connectivity}}, \code{\link{neuprint_get_adjacency_matrix}}
}
