% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{neuprint_skeleton_connectivity_matrix}
\alias{neuprint_skeleton_connectivity_matrix}
\title{Get a matrix for connectivity between neuron/neuronlist objects}
\usage{
neuprint_skeleton_connectivity_matrix(pre, post = NULL, ...)
}
\arguments{
\item{pre}{a neuron/neuronlist object. Putative input neurons (rows of returned matrix). If \code{post} is \code{NULL}, then these are also the putative target neurons (columns of returned matrix)}

\item{post}{a neuron/neuronlist object. Putative target neurons. Defaults to \code{NULL}}

\item{...}{methods sent to \code{nat::nlapply}}
}
\value{
a n x n matrix, where the rows are input neurons and the columns are their targets. Names are bodyids.
}
\description{
Get an adjacency matrix for the synaptic connectivity between \code{nat::neuron}/\code{nat::neuronlist} objects. This function does not query a neuPrint server.
It uses information on synaptic connectivity stored in a \code{nat::neuron}/\code{nat::neuronlist} object, as read from neuPrint by \code{neuprint_read_neurons}.
This can be particularly useful if you have neurons that you have been pruned using \code{nat:prune} family functions, because you just want to know the connectivity associated with
this modified skeleton, and not all connectivity associated with a bodyid on neuPrint.
}
\examples{
\donttest{
neurons = neuprint_read_neurons(c(818983130,1143677310))
M = neuprint_skeleton_connectivity_matrix(neurons)
stats::heatmap(M)
}
}
\seealso{
\code{\link{neuprint_connection_table}}, \code{\link{neuprint_read_neurons}}, \code{\link{neuprint_common_connectivity}}
}
