% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapses.R
\name{neuprint_synapse_connections}
\alias{neuprint_synapse_connections}
\title{Get the IDs of the pre- and post- synapses that make up a synaptic object}
\usage{
neuprint_synapse_connections(connector_ids, prepost = c(0, 1))
}
\arguments{
\item{prepost}{whether the given ID corresponds to a pre- (0) or post (1) synapse.}

\item{connectorID}{the ID of a synaptic connection object.}
}
\value{
a data frame, where each entry gives the pre-post associations for a synapse.
}
\description{
Get synapse IDs for a given connectorID, as well as their associated bodyIDs.
}
\examples{
\donttest{
syns = neuprint_get_synapses(818983130) # note
# all given connectorIDs from this function are for presynapses
prepost.conn = neuprint_synapse_connections(syns$connector_id, prepost = 0)
}
}
\seealso{
\code{\link{neuprint_fetch_custom}},
  \code{\link{neuprint_get_synapses}}
}
