% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.R
\name{neuprint_cosine_matrix}
\alias{neuprint_cosine_matrix}
\title{Calculate a cosine similarity matrix for neuprint neurons}
\usage{
neuprint_cosine_matrix(
  ids,
  ...,
  threshold = 5,
  partners = c("outputs", "inputs"),
  group = FALSE,
  groupfun = NULL,
  details = NULL,
  conn = NULL
)
}
\arguments{
\item{ids}{Passed to \code{\link{neuprint_ids}}}

\item{...}{Optional filter expression defining which partners to include}

\item{threshold}{An integer threshold (connections >= this will be returned)}

\item{partners}{Whether to cluster based on connections to input or output
partner neurons (default both).}

\item{group}{Whether to group by cell \code{type} or another named column.}

\item{groupfun}{A function which receives the metadata for all partner
neurons and returns a single grouping vector (see the \bold{details}
section).}

\item{details}{Optional character vector naming metadata columns to fetch for
partner neurons.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}
}
\value{
matrix or list of two matrices (input and output)
}
\description{
Calculate a cosine similarity matrix for neuprint neurons
}
\details{
For most purposes you can use \code{\link{neuprint_cosine_plot}}
  directly, but it can sometimes be useful to use
  \code{neuprint_cosine_matrix} to have more control over how partner neurons
  are grouped (see e.g. \code{groupfun}) or which partner neurons are
  included in the results (.

  The \code{groupfun} argument can be a powerful way to construct flexible
  grouping strategies for partner neurons. It was added in order to use
  information present in fields such as the group, serial or instance/name
  columns in the male VNC/CNS datasets. It will receive as input a dataframe
  and expects to receive a single vector of length matching the number of
  rows in the input dataframe. The input dataframe will contain the standard
  columns returned by \code{\link{neuprint_connection_table}} but you can
  request extra columns if necessary by naming them in the \code{group}
  argument.
}
\examples{
\donttest{
# NB the second (unnamed argument) filters the partner neurons
# so that only those with type containing the regular expression ORN are used
neuprint_cosine_matrix("/DA[1-3].*PN", grepl("ORN",type), partners='in')
}
}
\seealso{
\code{\link{neuprint_cosine_plot}}
}
