% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher.R
\name{neuprint_fetch_custom}
\alias{neuprint_fetch_custom}
\title{Use a cypher to make a custom query to the neuPrint server specifying
  the information you want to obtain}
\usage{
neuprint_fetch_custom(
  cypher = "MATCH (n:Meta) RETURN n.dataset, n.meshHost",
  dataset = NULL,
  conn = NULL,
  cache = FALSE,
  ...
)
}
\arguments{
\item{cypher}{the cypher by which to make your search, the default returns
the available datasets and the servers that host their associated mesh data}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link{neuprint_login}} for details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}

\item{cache}{if \code{TRUE} will use memoisation to cache the result of the
call for 1 hour.}

\item{...}{methods passed to \code{neuprint_login}}
}
\description{
Get summary information about the datasets hosted by the
  neuPrint server in which you are interested
}
\seealso{
\code{\link{neuprint_login}}, \code{\link{neuprint_available}}
}
