% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{neuprint_get_adjacency_matrix}
\alias{neuprint_get_adjacency_matrix}
\title{Get a matrix of connectivities between bodies}
\usage{
neuprint_get_adjacency_matrix(
  bodyids = NULL,
  inputids = NULL,
  outputids = NULL,
  dataset = NULL,
  all_segments = FALSE,
  conn = NULL,
  ...
)
}
\arguments{
\item{bodyids}{the body IDs for neurons/segments (bodies) you wish to query.
This can be in any form understood by \code{\link{neuprint_ids}}.}

\item{inputids, outputids}{identifiers for input and output bodies (use as an
alternative to \code{bodyids})}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link{neuprint_login}} for details.}

\item{all_segments}{if TRUE, all bodies are considered, if FALSE, only
'Neurons', i.e. bodies with a status roughly traced status.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}

\item{...}{methods passed to \code{neuprint_login}}
}
\value{
a n x n matrix, where the rows are input neurons and the columns are
  their targets. Only neurons supplied as the argument `bodyids` are
  considered.
}
\description{
Get an adjacency matrix for the synaptic connectivity within a
  set of specified bodies
}
\examples{
\donttest{
# these will mostly be axo-axonic connections between DA2 PNs
neuprint_get_adjacency_matrix('DA2 lPN')

# rectangular matrix with different in/out neurons
neuprint_get_adjacency_matrix(inputids='DA2 lPN', outputids='DL4 adPN')
}
\dontrun{
pnkc=neuprint_get_adjacency_matrix(inputids='name:mPN', outputids='/KC.*')
hist(colSums(pnkc), xlab = 'PN inputs / KC', br=100)
sum(rowSums(pnkc)>0)
}
}
\seealso{
\code{\link{neuprint_fetch_custom}},
  \code{\link{neuprint_simple_connectivity}},
  \code{\link{neuprint_common_connectivity}}
}
