% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{neuprint_get_adjacency_matrix}
\alias{neuprint_get_adjacency_matrix}
\title{Get an matrix of connectivities between neurons}
\usage{
neuprint_get_adjacency_matrix(bodyids, dataset = NULL, conn = NULL,
  ...)
}
\arguments{
\item{bodyids}{the body IDs for neurons/segments (bodies) you wish to query}

\item{dataset}{optional, a dataset you want to query. If NULL, the default specified by your R environ file is used. See \code{neuprint_login} for details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the neuPrint server see \code{?neuprint_login}.
If NULL, your defaults set in your R.profile or R.environ are used.}

\item{...}{methods passed to \code{neuprint_login}}
}
\value{
a n x n matrix, where the rows are input neurons and the columns are their targets
}
\description{
Get an adjacency matrix for the synaptic connectivity within a set of specified bodies
}
\seealso{
\code{\link{neuprint_fetch_custom}}, \code{\link{neuprint_simple_connectivity}}, \code{\link{neuprint_common_connectivity}}
}
