% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{neuprint_get_meta}
\alias{neuprint_get_meta}
\title{Get meta data associated with a body}
\usage{
neuprint_get_meta(
  bodyids,
  dataset = NULL,
  all_segments = TRUE,
  conn = NULL,
  ...
)
}
\arguments{
\item{bodyids}{the body IDs for neurons/segments (bodies) you wish to query}

\item{dataset}{optional, a dataset you want to query. If NULL, the default specified by your R environ file is used. See \code{neuprint_login} for details.}

\item{all_segments}{if TRUE, all bodies are considered, if FALSE, only 'Neurons', i.e. bodies with a status roughly traced status.}

\item{conn}{optional, a neuprintr connection object, which also specifies the neuPrint server see \code{?neuprint_login}.
If NULL, your defaults set in your R.profile or R.environ are used.}

\item{...}{methods passed to \code{neuprint_login}}
}
\value{
a dataframe, one row for each given body id, columns bodyid, name, status, voxels, pre and post. If data is missing, NA is returned.
}
\description{
Return important meta data given bodyids, including the putative neuron's name, status, size in voxels, and number of pre and post synapses.
}
\examples{
\donttest{
neuprint_get_meta(c(818983130, 1796818119))
}
}
