% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{neuprint_get_meta}
\alias{neuprint_get_meta}
\title{Get key metadata for body (including name, type, status, size)}
\usage{
neuprint_get_meta(
  bodyids,
  dataset = NULL,
  all_segments = TRUE,
  conn = NULL,
  ...
)
}
\arguments{
\item{bodyids}{the body IDs for neurons/segments (bodies) you wish to query}

\item{dataset}{optional, a dataset you want to query. If NULL, the default specified by your R environ file is used. See \code{neuprint_login} for details.}

\item{all_segments}{if TRUE, all bodies are considered, if FALSE, only 'Neurons', i.e. bodies with a status roughly traced status.}

\item{conn}{optional, a neuprintr connection object, which also specifies the neuPrint server see \code{?neuprint_login}.
If NULL, your defaults set in your R.profile or R.environ are used.}

\item{...}{methods passed to \code{neuprint_login}}
}
\value{
a \code{data.frame} containing the neuron's \itemize{ \item name

  \item type Cell type of the neuron

  \item status (Traced etc)

  \item statusLabel similar to \code{status} but often a bit more specific
  \item size size in voxels

  \item pre number of presynapses

  \item post number of postsynapses

  \item cropped whether the neuron is cropped by the hemibrain volume

  \item soma whether the neuron has a soma in the hemibrain volume

  \item cellBodyFiber names the tract connecting the soma to rest of neuron

  }
}
\description{
Get key metadata for body (including name, type, status, size)
}
\details{
Sometimes a \code{cellBodyFiber} will be present even when the
  \code{soma} is not, so this may be a good test of if the neuron is present
  in the volume. The \code{cellBodyFiber} should be matched to (hemi)lineage
  information .
}
\examples{
\donttest{
da2s=neuprint_search(".*DA2.*")
neuprint_get_meta(da2s$bodyid)
}
}
