% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{neuprint_get_roiInfo}
\alias{neuprint_get_roiInfo}
\title{Get roiInfo associated with a body}
\usage{
neuprint_get_roiInfo(
  bodyids,
  dataset = NULL,
  all_segments = FALSE,
  conn = NULL,
  ...
)
}
\arguments{
\item{bodyids}{the body IDs for neurons/segments (bodies) you wish to query.
This can be in any form understood by \code{\link{neuprint_ids}}.}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link{neuprint_login}} for details.}

\item{all_segments}{if TRUE, all bodies are considered, if FALSE, only
'Neurons', i.e. bodies with a status roughly traced status.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}

\item{...}{methods passed to \code{neuprint_login}}
}
\value{
a dataframe, one row for each given body id, columns ROI_pre and ROI_post for every ROI. If data is missing, NA is returned.
}
\description{
Return pre and post counts in all the ROIs given bodyids innervate.
}
\examples{
\donttest{
neuprint_get_roiInfo(c(818983130, 1796818119))
}
}
