% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapses.R
\name{neuprint_get_synapses}
\alias{neuprint_get_synapses}
\title{Get the locations of pre- and post synapses associated with a body}
\usage{
neuprint_get_synapses(bodyid = bodyid, dataset = NULL, roi = "all",
  conn = NULL, ...)
}
\arguments{
\item{bodyid}{the body ID for a neuron/segment (body) you wish to query}

\item{dataset}{optional, a dataset you want to query. If NULL, the default specified by your R environ file is used. See \code{neuprint_login} for details.}

\item{roi}{a roi (i.e. neuropil volume) you want to query. Use \code{neuprint_ROIs} to see what is available.
Defaults to 'all', which will return synapses in all ROIs.}

\item{conn}{optional, a neuprintr connection object, which also specifies the neuPrint server see \code{?neuprint_login}.
If NULL, your defaults set in your R.profile or R.environ are used.}

\item{...}{methods passed to \code{neuprint_login}}
}
\value{
a n x n matrix, where the rows are input neurons and the columns are their targets
}
\description{
Get the xyz locations, IDs and confidences of pre- and post synapses associated with a body
}
\seealso{
\code{\link{neuprint_fetch_custom}}, \code{\link{neuprint_simple_connectivity}}, \code{\link{neuprint_common_connectivity}}
}
