% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapses.R
\name{neuprint_get_synapses}
\alias{neuprint_get_synapses}
\title{Get the locations of pre- and post synapses associated with a body.}
\usage{
neuprint_get_synapses(
  bodyids,
  roi = NULL,
  remove.autapses = TRUE,
  chunk = TRUE,
  progress = NULL,
  dataset = NULL,
  conn = NULL,
  ...
)
}
\arguments{
\item{bodyids}{the body IDs for neurons/segments (bodies) you wish to query.
This can be in any form understood by \code{\link{neuprint_ids}}.}

\item{roi}{a ROI (i.e. neuropil volume) you want to query. Use
\code{\link{neuprint_ROIs}} to see what is available. Defaults to 'all',
which will return synapses in all ROIs.}

\item{remove.autapses}{Whether to remove autaptic connections (default TRUE)
from the results. It appears that there is a relatively high number of
false positive autapses.}

\item{chunk}{A logical specifying whether to split the query into multiple
chunks or an integer specifying the size of those chunks (which defaults to
20 when \code{chunk=TRUE}).}

\item{progress}{if TRUE, a progress bar will be shown. This may slow the data
fetching process for smaller queries (<100 body ids). The default of
\code{progress=NULL} will only show a progress bar if the query will be
split into multiple chunks based on the \code{chunk} argument.}

\item{dataset}{optional, a dataset you want to query. If NULL, the default
specified by your R environ file is used. See \code{\link{neuprint_login}}
for details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}

\item{...}{methods passed to \code{neuprint_login}}
}
\value{
a data frame, where each row is a connection between one query
  \code{bodyid} and a single \code{partner} neuron. The \code{prepost} column
  should be interpreted as follows: \describe{

  \item{\code{prepost=0}}{ a downstream or \emph{output} partner,
  postsynaptic to the query neuron}

  \item{\code{prepost=1}}{ an upstream or \emph{input} partner, presynaptic
  to the query neuron}

  } Each connection is associated with a synapse that has its own unique
  \code{connector_id}, XYZ coordinates and a confidence in its existence.
}
\description{
Get the xyz locations, IDs and confidences of pre- and post
  synapses associated with a body. The \code{connector_id} for each synapse
  is for the 'presynaptic' object. Separate IDs exist for the postsynapses in
  neuPrint, and these are not fetched b this function. This is similar to the
  logic use in \code{rcatmaid}.
}
\examples{
\donttest{
syns=neuprint_get_synapses(c(818983130, 1796818119))
head(syns)
}
}
\seealso{
\code{\link{neuprint_fetch_custom}},
  \code{\link{neuprint_simple_connectivity}},
  \code{\link{neuprint_common_connectivity}}
}
