% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{neuprint_datasets}
\alias{neuprint_datasets}
\alias{neuprint_database}
\alias{neuprint_available}
\alias{neuprint_version}
\title{Find out what some information about your neuPrint server}
\usage{
neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)
}
\arguments{
\item{...}{methods passed to \code{neuprint_login}}
}
\description{
Get summary information about the datasets hosted by the neuPrint server in which you are interested
}
\examples{
\donttest{
# list details for all the available datasets
neuprint_datasets()
}
\donttest{
# list API endpoints
head(neuprint_available())
}
\donttest{
neuprint_version()
}
}
\seealso{
\code{\link{neuprint_login}}
}
