% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{neuprint_ROI_connectivity}
\alias{neuprint_ROI_connectivity}
\title{Get the connectivity between ROIs in a neuPrint dataset}
\usage{
neuprint_ROI_connectivity(
  rois,
  cached = FALSE,
  full = TRUE,
  statistic = c("weight", "count"),
  dataset = NULL,
  conn = NULL,
  ...
)
}
\arguments{
\item{rois}{regions of interest for a dataset}

\item{cached}{pull cached results (TRUE) or recalculate the connectivity (FALSE)?}

\item{full}{return all neurons involved (TRUE, the default) or give a ROI summary (FALSE, default behavior if `cached` is TRUE)}

\item{statistic}{either "weight" or count" (default "weight"). Which number to return (see neuprint explorer for details) for summary results (either `full` is FALSE or `cached` is TRUE)}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link{neuprint_login}} for details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link{neuprint_login}} for details.}

\item{...}{methods passed to \code{neuprint_login}}
}
\description{
Get the connectivity between ROIs in a neuPrint dataset
}
\seealso{
\code{\link{neuprint_simple_connectivity}}, \code{\link{neuprint_common_connectivity}}
}
