% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{neuprint_ROI_hierarchy}
\alias{neuprint_ROI_hierarchy}
\title{Get the region of interest (ROI) hierarchy in a dataset}
\usage{
neuprint_ROI_hierarchy(dataset = NULL, conn = NULL, ...)
}
\arguments{
\item{dataset}{optional, a dataset you want to query. If NULL, the default
specified by your R environ file is used. See \code{neuprint_login} for
details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server see \code{\link{neuprint_login}}. If NULL, your defaults
set in your R.profile or R.environ are used.}

\item{...}{methods passed to \code{neuprint_fetch_custom}}
}
\description{
Get a data frame describing how ROIs are related.
}
\examples{
\donttest{
roi.hierarchy = neuprint_ROI_hierarchy()
g = igraph::graph_from_data_frame(roi.hierarchy, directed = TRUE)
igraph::plot.igraph(g, layout=igraph::layout_as_tree)
}
}
\seealso{
\code{\link{neuprint_ROIs}}, \code{\link{neuprint_get_roiInfo}}
}
