% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{neuprint_ROI_mesh}
\alias{neuprint_ROI_mesh}
\title{Import a region of interest as a mesh}
\usage{
neuprint_ROI_mesh(roi, dataset = NULL, conn = NULL, ...)
}
\arguments{
\item{roi}{region of interest for a dataset}

\item{dataset}{optional, a dataset you want to query. If NULL, the default specified by your R environ file is used. See \code{neuprint_login} for details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the neuPrint server see \code{?neuprint_login}.
If NULL, your defaults set in your R.profile or R.environ are used.}

\item{...}{methods passed to \code{neuprint_login}}
}
\description{
Import a region of interest as a mesh
}
