% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{neuprint_ROIs}
\alias{neuprint_ROIs}
\title{Get a vector of all the ROIs in a specified dataset}
\usage{
neuprint_ROIs(superLevel = FALSE, dataset = NULL, conn = NULL, ...)
}
\arguments{
\item{dataset}{optional, a dataset you want to query. If NULL, the default
specified by your R environ file is used. See \code{neuprint_login} for
details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server see \code{\link{neuprint_login}}. If NULL, your defaults
set in your R.profile or R.environ are used.}

\item{...}{methods passed to \code{neuprint_login}}

\item{superlevel}{whether not to show 'superlevel' ROIs - ROIs composed of other ROIs.
If set to NULL, both low-level and superlevel ROIs are returned.}
}
\description{
Get the regions of interest (ROIs) used in a neuPrint project
}
\seealso{
\code{\link{neuprint_login}}, \code{\link{neuprint_datasets}}
}
