% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprintr-package.R
\docType{package}
\name{neuprintr-package}
\alias{neuprintr-package}
\alias{_PACKAGE}
\alias{neuprintr}
\title{neuprintr: R client utilities for interacting with the neuPrint connectome analysis service}
\description{
Access to the neuprint connectommie analysis service, including 
    functions to fetch skeletons, synapses and ROIs, query connnectivity data
    and develop new low level queries.
}
\section{Online Documentation}{
 See
  \url{https://jefferislab.github.io/neuprintr/} for structured online
  documentation.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferislab/neuprintr}
  \item Report bugs at \url{https://github.com/jefferislab/neuprintr/issues}
}

}
\keyword{internal}
