% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprintr-package.R
\docType{package}
\name{neuprintr-package}
\alias{neuprintr-package}
\alias{_PACKAGE}
\alias{neuprintr}
\title{neuprintr: R client utilities for interacting with the neuPrint connectome analysis service}
\description{
Access to the neuprint connectome analysis service, including 
    functions to fetch skeletons, synapses and ROIs, query connectivity data
    and develop new low level queries.
}
\section{Login}{
 See \code{\link{neuprint_login}} or the package README for
  details of the how to login to a neuPrint server, the recommended initial
  setup and (optionally) how to specify a default dataset.
}

\section{Online Documentation}{
 See
  \url{https://natverse.github.io/neuprintr/} for structured online
  documentation.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/neuprintr}
  \item Report bugs at \url{https://github.com/natverse/neuprintr/issues}
}

}
\author{
\strong{Maintainer}: Alexander Bates \email{alexander.shakeel.bates@gmail.com} (\href{https://orcid.org/0000-0002-1195-0445}{ORCID})

Authors:
\itemize{
  \item Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})
  \item Romain Franconville (\href{https://orcid.org/0000-0002-4440-7297}{ORCID})
}

}
\keyword{internal}
