% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprintr-package.R
\docType{package}
\name{neuprintr-package}
\alias{neuprintr-package}
\alias{neuprintr}
\title{neuprintr: R Client Utilities for Interacting with the neuPrint Connectome Analysis Service}
\description{
Access to the neuprint connectome analysis service, including functions to fetch skeletons, synapses and ROIs, query connectivity data and develop new low level queries.
}
\section{Login}{
 See \code{\link{neuprint_login}} or the package README for
  details of the how to login to a neuPrint server, the recommended initial
  setup and (optionally) how to specify a default dataset.
}

\section{Online Documentation}{
 See
  \url{https://natverse.github.io/neuprintr/} for structured online
  documentation.
}

\section{Package Options}{
 \itemize{

  \item \code{neuprintr.support_hidden_rois} In some instances rois are not
  defined in a dataset's ROI list but only in field names. However, in these
   circumstances, these ROI fields will not be listed in the neuprint field
   properties for the dataset and a slower approach must be used to list
   fields. To avoid paying a speed penalty for what is a rare circumstance
   even with private in progress datasets, this option is provided. Set it to
   \code{TRUE} to support such hidden fields if you need them.
  }
}

\examples{
# List all package options
\dontrun{
options()[grepl("^neuprintr", names(options()))]
}

}
\seealso{
Useful links:
\itemize{
  \item \url{https://natverse.org/neuprintr}
  \item \url{https://github.com/natverse/neuprintr}
  \item Report bugs at \url{https://github.com/natverse/neuprintr/issues}
}

}
\author{
\strong{Maintainer}: Alexander Bates \email{alexander.shakeel.bates@gmail.com} (\href{https://orcid.org/0000-0002-1195-0445}{ORCID})

Authors:
\itemize{
  \item Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})
  \item Romain Franconville (\href{https://orcid.org/0000-0002-4440-7297}{ORCID})
}

}
\keyword{internal}
