% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprintr-package.R
\docType{package}
\name{neuprintr-package}
\alias{neuprintr-package}
\alias{_PACKAGE}
\alias{neuprintr}
\title{neuprintr: R client utilities for interacting with the neuPrint connectome analysis service}
\description{
Access to the neuprint connectome analysis service, including 
    functions to fetch skeletons, synapses and ROIs, query connectivity data
    and develop new low level queries.
}
\section{Online Documentation}{
 See
  \url{https://natverse.github.io/neuprintr/} for structured online
  documentation.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/neuprintr}
  \item Report bugs at \url{https://github.com/natverse/neuprintr/issues}
}

}
\author{
\strong{Maintainer}: Alexander Bates \email{alexander.shakeel.bates@gmail.com} (0000-0002-1195-0445)

Authors:
\itemize{
  \item Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)
}

Other contributors:
\itemize{
  \item Romain Franconville (0000-0002-4440-7297) [contributor]
}

}
\keyword{internal}
